# backend/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"ErrorHttpResponse": {
				"description": "Error response model",
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"details": {
						"description": "Error details",
						"example": {}
					}
				},
				"required": [
					"httpCode",
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiTokensResponse": {
				"description": "API Token list response model",
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"apiTokens": {
						"items": {
							"$ref": "#/components/schemas/ApiTokenHttpEntity"
						},
						"type": "array",
						"description": "API token list entity"
					},
					"count": {
						"type": "number",
						"format": "double",
						"description": "API token count"
					}
				},
				"required": [
					"httpCode",
					"error",
					"message",
					"apiTokens",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiTokenCreateResponse": {
				"description": "API Token creation response model",
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"apiToken": {
						"$ref": "#/components/schemas/ApiTokenHttpEntity",
						"description": "API token entity"
					},
					"bearerToken": {
						"type": "string",
						"description": "Bearer token to login with api key\n  // eslint-disable-next-line max-len",
						"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
					}
				},
				"required": [
					"httpCode",
					"error",
					"message",
					"apiToken",
					"bearerToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiTokenCreateHttpEntity": {
				"description": "API Token creation entity model",
				"properties": {
					"name": {
						"type": "string",
						"description": "Token name",
						"example": "API key for CI",
						"pattern": ".*[\\S\\s].*"
					},
					"description": {
						"type": "string",
						"nullable": true,
						"description": "Token description (optional)",
						"example": "This is a description"
					},
					"expireAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"description": "Token expiration date (optional)",
						"example": "2023-01-01T00:00:00.000Z"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Color": {
				"type": "string",
				"description": "Hexadecimal color",
				"pattern": "#[0-9A-Za-z]{6}"
			},
...
			"BadgesHttpResponse": {
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"badges": {
						"items": {
							"$ref": "#/components/schemas/BadgeHttpEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"httpCode",
					"error",
					"message",
					"count",
					"badges"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BadgeHttpResponse": {
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"badge": {
						"$ref": "#/components/schemas/BadgeHttpEntity"
					}
				},
				"required": [
					"httpCode",
					"error",
					"message",
					"badge"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BadgeCreateUpdateHttpEntity": {
				"description": "Badge create/update entity model",
				"properties": {
					"name": {
						"type": "string",
						"description": "New badge name",
						"example": "Badge",
						"pattern": ".*[\\S\\s].*"
					},
					"color": {
						"$ref": "#/components/schemas/Color",
						"description": "New badge color",
						"example": "#1A2B3C"
					},
					"description": {
						"type": "string",
						"nullable": true,
						"description": "New badge description (optional)",
						"example": "This is a description"
					}
				},
				"required": [
					"name",
					"color"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FilesHttpResponse": {
				"description": "File list response model",
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"count": {
						"type": "number",
						"format": "double",
						"description": "Stored files count"
					},
					"files": {
						"items": {
							"$ref": "#/components/schemas/FileHttpEntity"
						},
						"type": "array",
						"description": "File list"
					}
				},
				"required": [
					"httpCode",
					"error",
					"message",
					"count",
					"files"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GroupsHttpResponse": {
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"count": {
						"type": "number",
						"format": "double"
					},
					"groups": {
						"items": {
							"$ref": "#/components/schemas/GroupHttpEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"httpCode",
					"error",
					"message",
					"count",
					"groups"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GroupHttpResponse": {
				"properties": {
					"httpCode": {
						"type": "number",
						"format": "double",
						"description": "Response HTTP code",
						"example": "200"
					},
					"error": {
						"type": "boolean",
						"description": "Error status",
						"example": "false"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message of response",
						"example": "Sample message"
					},
					"group": {
						"$ref": "#/components/schemas/GroupHttpEntity"
					}
				},
				"required": [
					"httpCode",
					"error",
					"message",
					"group"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Use /login endpoint to get bearer token or use an API token"
			}
		}
	},
	"info": {
		"title": "open-artifactory",
		"version": "1.0.0",
		"description": "Open source artifactory",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "https://github.com/3jedgcm/open-artifactory#author"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/tokens": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "API token list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiTokensResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 200,
											"error": false,
											"message": "1 api token found",
											"count": 1,
											"apiTokens": [
												{
													"id": 1,
													"name": "API key for CI",
													"description": "This is a description",
													"expireAt": "2023-01-01T00:00:00.000Z",
													"isValid": true,
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z"
												}
											]
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 401,
											"message": "Unauthorized",
											"error": true
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 500,
											"message": "Internal server error",
											"error": true,
											"details": {
												"message": "Sample error"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Gets API token metadata list",
				"summary": "Gets API token list",
				"tags": [
					"Tokens"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Generate",
				"responses": {
					"201": {
						"description": "Created API token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiTokenCreateResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 200,
											"error": false,
											"message": "#1 api token generated",
											"apiToken": {
												"id": 1,
												"name": "API key for CI",
												"description": "This is a description",
												"expireAt": "2023-01-01T00:00:00.000Z",
												"isValid": true,
												"createdAt": "2023-01-01T00:00:00.000Z",
												"updateAt": "2023-01-01T00:00:00.000Z"
											},
											"bearerToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 401,
											"message": "Unauthorized",
											"error": true
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation failed error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 422,
											"message": "Validation failed",
											"error": true,
											"details": {
												"field": {
													"message": "Validation message",
													"value": "Bad value"
												}
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 500,
											"message": "Internal server error",
											"error": true,
											"details": {
												"message": "Sample error"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Generates API token",
				"summary": "Generates API token",
				"tags": [
					"Tokens"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "token generation parameters",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ApiTokenCreateHttpEntity",
								"description": "token generation parameters"
							}
						}
					}
				}
			}
		},
...
		"/badges": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Badge list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadgesHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 200,
											"error": false,
											"message": "3 badges found",
											"count": 3,
											"badges": [
												{
													"id": 1,
													"name": "Group",
													"color": "#FFFFFF",
													"description": "This is a description",
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z"
												},
												{
													"id": 1,
													"name": "Group",
													"color": "#FFFFFF",
													"description": "This is a description",
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z"
												},
												{
													"id": 1,
													"name": "Group",
													"color": "#FFFFFF",
													"description": "This is a description",
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z"
												}
											]
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 401,
											"message": "Unauthorized",
											"error": true
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 500,
											"message": "Internal server error",
											"error": true,
											"details": {
												"message": "Sample error"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Gets badges list",
				"summary": "Gets badges",
				"tags": [
					"Badges"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Created badge",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadgeHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 200,
											"error": false,
											"message": "#1 badge created",
											"badge": {
												"id": 1,
												"name": "Group",
												"color": "#FFFFFF",
												"description": "This is a description",
												"createdAt": "2023-01-01T00:00:00.000Z",
												"updateAt": "2023-01-01T00:00:00.000Z"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 401,
											"message": "Unauthorized",
											"error": true
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation failed error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 422,
											"message": "Validation failed",
											"error": true,
											"details": {
												"field": {
													"message": "Validation message",
													"value": "Bad value"
												}
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 500,
											"message": "Internal server error",
											"error": true,
											"details": {
												"message": "Sample error"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates badge",
				"summary": "Creates badge",
				"tags": [
					"Badges"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "badge data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BadgeCreateUpdateHttpEntity",
								"description": "badge data"
							}
						}
					}
				}
			}
		},
...
		"/files": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "File list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FilesHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 200,
											"error": false,
											"message": "3 stored files",
											"count": 3,
											"files": [
												{
													"uuid": "146fb209-af3b-4c67-863a-a98b641c95e5",
													"name": "picture.jpg",
													"comment": "This is a comment",
													"mimeType": "image/jpeg",
													"size": 10578,
													"hash": "7f4fd6614997f869edf8578d81ceba9f",
													"downloadCount": 10,
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z",
													"url": "https://exemple.com/files/146fb209-af3b-4c67-863a-a98b641c95e5",
													"group": {
														"id": 1,
														"name": "Group",
														"description": "This is a description",
														"createdAt": "2023-01-01T00:00:00.000Z",
														"updateAt": "2023-01-01T00:00:00.000Z"
													},
													"badges": [
														{
															"id": 1,
															"name": "Group",
															"color": "#FFFFFF",
															"description": "This is a description",
															"createdAt": "2023-01-01T00:00:00.000Z",
															"updateAt": "2023-01-01T00:00:00.000Z"
														},
														{
															"id": 1,
															"name": "Group",
															"color": "#FFFFFF",
															"description": "This is a description",
															"createdAt": "2023-01-01T00:00:00.000Z",
															"updateAt": "2023-01-01T00:00:00.000Z"
														}
													]
												},
												{
													"uuid": "146fb209-af3b-4c67-863a-a98b641c95e5",
													"name": "picture.jpg",
													"comment": "This is a comment",
													"mimeType": "image/jpeg",
													"size": 10578,
													"hash": "7f4fd6614997f869edf8578d81ceba9f",
													"downloadCount": 10,
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z",
													"url": "https://exemple.com/files/146fb209-af3b-4c67-863a-a98b641c95e5",
													"group": {
														"id": 1,
														"name": "Group",
														"description": "This is a description",
														"createdAt": "2023-01-01T00:00:00.000Z",
														"updateAt": "2023-01-01T00:00:00.000Z"
													},
													"badges": [
														{
															"id": 1,
															"name": "Group",
															"color": "#FFFFFF",
															"description": "This is a description",
															"createdAt": "2023-01-01T00:00:00.000Z",
															"updateAt": "2023-01-01T00:00:00.000Z"
														},
														{
															"id": 1,
															"name": "Group",
															"color": "#FFFFFF",
															"description": "This is a description",
															"createdAt": "2023-01-01T00:00:00.000Z",
															"updateAt": "2023-01-01T00:00:00.000Z"
														}
													]
												},
												{
													"uuid": "146fb209-af3b-4c67-863a-a98b641c95e5",
													"name": "picture.jpg",
													"comment": "This is a comment",
													"mimeType": "image/jpeg",
													"size": 10578,
													"hash": "7f4fd6614997f869edf8578d81ceba9f",
													"downloadCount": 10,
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z",
													"url": "https://exemple.com/files/146fb209-af3b-4c67-863a-a98b641c95e5",
													"group": {
														"id": 1,
														"name": "Group",
														"description": "This is a description",
														"createdAt": "2023-01-01T00:00:00.000Z",
														"updateAt": "2023-01-01T00:00:00.000Z"
													},
													"badges": [
														{
															"id": 1,
															"name": "Group",
															"color": "#FFFFFF",
															"description": "This is a description",
															"createdAt": "2023-01-01T00:00:00.000Z",
															"updateAt": "2023-01-01T00:00:00.000Z"
														},
														{
															"id": 1,
															"name": "Group",
															"color": "#FFFFFF",
															"description": "This is a description",
															"createdAt": "2023-01-01T00:00:00.000Z",
															"updateAt": "2023-01-01T00:00:00.000Z"
														}
													]
												}
											]
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 401,
											"message": "Unauthorized",
											"error": true
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 500,
											"message": "Internal server error",
											"error": true,
											"details": {
												"message": "Sample error"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Gets file list",
				"summary": "Gets files",
				"tags": [
					"Files"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
...
		"/groups": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Group list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupsHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 200,
											"error": false,
											"message": "3 groups found",
											"count": 3,
											"groups": [
												{
													"id": 1,
													"name": "Group",
													"description": "This is a description",
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z"
												},
												{
													"id": 1,
													"name": "Group",
													"description": "This is a description",
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z"
												},
												{
													"id": 1,
													"name": "Group",
													"description": "This is a description",
													"createdAt": "2023-01-01T00:00:00.000Z",
													"updateAt": "2023-01-01T00:00:00.000Z"
												}
											]
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 401,
											"message": "Unauthorized",
											"error": true
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 500,
											"message": "Internal server error",
											"error": true,
											"details": {
												"message": "Sample error"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Gets group list",
				"summary": "Gets groups",
				"tags": [
					"Groups"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Created group",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GroupHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 200,
											"error": false,
											"message": "#1 group created",
											"group": {
												"id": 1,
												"name": "Group",
												"description": "This is a description",
												"createdAt": "2023-01-01T00:00:00.000Z",
												"updateAt": "2023-01-01T00:00:00.000Z"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 401,
											"message": "Unauthorized",
											"error": true
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation failed error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 422,
											"message": "Validation failed",
											"error": true,
											"details": {
												"field": {
													"message": "Validation message",
													"value": "Bad value"
												}
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorHttpResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"httpCode": 500,
											"message": "Internal server error",
											"error": true,
											"details": {
												"message": "Sample error"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates group",
				"summary": "Creates group",
				"tags": [
					"Groups"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "group data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GroupCreateUpdateHttpEntity",
								"description": "group data"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5000/"
		}
	]
}